ml/requirements.txt
# Data science & ML
numpy
pandas
scikit-learn
joblib
scipy
nltk==3.9.1
matplotlib-inline==0.1.7

# Python environment & packaging
setuptools==80.9.0
packaging==25.0
typing_extensions==4.14.0
python-dotenv

# Misc tools
click==8.2.1
colorama==0.4.6
loguru==0.7.3
pillow==11.2.1
platformdirs==4.3.8
pycparser==2.22
six==1.17.0
timezonefinder
requests
urllib3
validators
beautifulsoup4
feedparser
regex

# Supabase
supabase

# Google APIs
google-api-python-client
google-auth
